requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.1

# Data Validation & Configuration  
pydantic==2.5.0
python-dotenv==1.0.0

# AI/ML Dependencies
openai==1.7.0
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.0
faiss-cpu==1.7.4

# Network & Security (SSL fix)
urllib3==1.26.18

# Text Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.8.1
tiktoken==0.5.2
python-docx==1.1.0

# Data Handling
numpy==1.26.2
pandas==2.1.4
tqdm==4.66.1

# Logging and Monitoring
loguru==0.7.2
psutil==5.9.6

# Performance & Async
aiofiles==23.2.0

# Security
cryptography==41.0.8

# Development Tools (optional - install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.12.0
# flake8==6.1.0
# mypy==1.8.0
# isort==5.13.2

# Production Server (optional)
gunicorn==21.2.0
uvicorn==0.24.0
waitress==2.1.2

# Logging
loguru>=0.7.2

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
